backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# RAG and LLM libraries
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.3.0

# Vector database (ChromaDB)
chromadb>=0.4.22

# LLM providers
google-generativeai==0.3.1

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data validation and models
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Development tools (linting and formatting)
black>=24.0.0
flake8>=7.0.0
isort>=5.13.0

